Add files via upload
House of 28 (previously known as KF Channel (S) Pte Ltd) is a multi-faceted company that retails products across 3 categories - fashion, beauty and home. The company manages both online and offline retail stores. The company's main objective is to help elevate many established and/or emerging fashion brands onto a global platform, making them more accessible to consumers around Asia, and the world.

House of 28 would like to determine various trends in each of the industries in order to cater to their customers better and adjust their inventory based on the demand. We were required to analyse data based on sales from 2019 to 2021 and to come up with insights and recommendations.